crypto: rewrite `check_sender_trust_requirement` to use `VerficationState`
There are two reasons for this.

Firstly. we've already done a bunch of work to map `SenderData` into a
`VerificationState`, and the decision tree from `VerificationState` to
allow/reject is simpler than going from `SenderData`, even if we have to fudge
it a bit to get the "legacy" flag. (Note that it allows us to get rid of an
`unreachable!` panic.)

Secondly, `VerficationState` represents the state of an *event*, whereas
`SenderData` is about the session as a whole. A session can be fine, whilst
event (claiming to be) encrypted with it can be suspect. What we want here
is to check a specific message. Currently, this doesn't make any functional
difference, but conceptually it's cleaner to check the `VerificationState`.