feat(ui): implement threaded message filtering on the `TimelineAction` level
This patch moves the event relation processing and thread root extraction to the timeline medatada so they can be used independent of content parsing.
It also passes the `timeline_focus_kind` and into the TimelineAction so it can be used to compute whether the current event/content is part of a thread or is the thread root itself.