multiverse: write logs to session dir
Rather than always writing the logs to `/tmp`, write the, to the session
directory. The session directory defaults to `/tmp` so by default this will do
the same as before, but if you override the session path on the commandline,
the logs will get stored alongside the stores and caches.

This is particularly useful when running two instances of multiverse, and you
want them to put their logs in different places.