fix(base): Remove false positives in `check_tombstone`.
This patch removes false positives in `check_tombstone`. See the new
`test_check_tombstone_no_loop_within_misordered_rooms` test which
explains the problem. In a nutshell, the `Response` type holds room
updates in a `HashMap`, thus the room updates are _re-ordered_ by the
`HashMap`. It creates problems with the algorithm in `check_tombstone`.
This patch re-orders room updates by room versions. Now we can have
false negatives if the room version is unknown. A discussion is
happening to make the `m.room.create` not optional in all the sync
mechanisms.