fix(timeline): take into account the skip count when returning the pagination status
The pagination status was only mapped onto the "global" state, that is,
the state of the event cache pagination.

Now, consider the following case, where two timeline instances are live:

- the first one could have backpaginated all the items back to the start
of the room
- the second one is created later. Because of the initial value for the
skip count, it will only return a subset of event items (~20).

However, listening to the pagination status for the second timeline
would incorrectly state that the timeline was entirely paginated
(because it returned the "global" pagination status). As such, an
observer might think that there are no more items in the timeline, while
a subsequent pagination would adjust the skip count and return more
items.

This fixes it by combining the global pagination state with the local
timeline state (aka the skip count value). If the skip count is positive
(meaning, we could set it to 0 later and thus returning more events), we
pretend we haven't reached the start of the timeline. This way, an
observer can call pagination later, which may adjust the skip count and
"return" more items.