Drop the historic room key bundles from the database as the data format changed
The format of the stored bundle changed as we now persist the Curve25519
key of the bundle sender.

Since this is still an experimental feature, and there's no way to get
to the Curve25519 key after the fact, we're going to drop all the stored
bundles.

Newly saved bundles will persist now the Curve25519 key as well.